+++ galpy/backend/_jax/__init__.py
@@ -1,0 +1,7 @@
+###############################################################################
+#   galpy.backend._jax: jax-specific backend implementations.
+#
+#   Code here imports jax / diffrax directly (it is genuinely jax-specific, as
+#   opposed to the data-first xp-dispatch compute layer). The torch counterparts
+#   live in galpy.backend._torch.
+###############################################################################
