+++ galpy/backend/_torch/__init__.py
@@ -1,0 +1,7 @@
+###############################################################################
+#   galpy.backend._torch: torch-specific backend implementations.
+#
+#   Code here imports torch / torchdiffeq directly (it is genuinely torch-
+#   specific, as opposed to the data-first xp-dispatch compute layer). The jax
+#   counterparts live in galpy.backend._jax.
+###############################################################################
